+++ src/components/buttons/styled-button.tsx
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+
+export const StyledButton = styled.div`
+  padding: 0.25em;
+  margin: 0.25em;
+  border: 1px solid hsl(0, 0%, 25%);
+  border-radius: 0.2em;
+`;
